+++ inst/doc/An_Introduction_to_packager.R
@@ -150,3 +150,9 @@ desc::desc_get("Version", file = path)
 cat(readLines(file.path(path, "NEWS.md")), sep = "\n")
 
 
+
+## ---- echo = FALSE------------------------------------------------------------
+package_title <- "veryImportantPackage"
+path <- file.path(tempdir(), package_title)
+unlink(path, recursive = TRUE)
+